[Pg] refactor: move PostGIS types into their own file
